+++ src/main/java/org/seasar/doma/wrapper/ShortWrapper.java
@@ -2,23 +2,13 @@ package org.seasar.doma.wrapper;
 
 import org.seasar.doma.DomaNullPointerException;
 
-/**
- * {@link Short} のラッパーです。
- *
- * @author taedium
- */
+/** A wrapper for the {@link Short} class. */
 public class ShortWrapper extends AbstractWrapper<Short> implements NumberWrapper<Short> {
 
-  /** インスタンスを構築します。 */
   public ShortWrapper() {
     super(Short.class);
   }
 
-  /**
-   * 値を指定してインスタンスを構築します。
-   *
-   * @param value 値
-   */
   public ShortWrapper(Short value) {
     super(Short.class, value);
   }
@@ -30,14 +20,14 @@ public class ShortWrapper extends AbstractWrapper<Short> implements NumberWrappe
 
   @Override
   public Short getDefault() {
-    return Short.valueOf((short) 0);
+    return (short) 0;
   }
 
   @Override
   public void increment() {
     Short value = doGet();
     if (value != null) {
-      doSet((short) (value.shortValue() + 1));
+      doSet((short) (value + 1));
     }
   }
 
@@ -45,7 +35,7 @@ public class ShortWrapper extends AbstractWrapper<Short> implements NumberWrappe
   public void decrement() {
     Short value = doGet();
     if (value != null) {
-      doSet((short) (value.shortValue() - 1));
+      doSet((short) (value - 1));
     }
   }
